chatbot/requirements.txt
Flask>=2.0
Flask-CORS>=3.0
python-dotenv>=0.19
requests>=2.25